pyproject.toml

[tool.poetry]
name = "paid-media-userlist"
version = "0.1.0"
description = ""
authors = ["Martin Højland <molheh@nuuday.dk>"]

[tool.poetry.dependencies]
python = "^3.9"
sqlalchemy = "^1.3.20"
tqdm = "^4.52.0"
pandas = "^1.1.4"
numpy = "^1.19.4"
mysqlclient = "^2.0.1"
requests = "^2.25.0"
pydantic = "^1.7.2"
python-dotenv = "^0.15.0"
xmltodict = "^0.12.0"
boto3 = "^1.16.25"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
black = {version = "^20.8b1", allow-prereleases = true}
pylint = "^2.6.0"
pytest-cov = "^2.10.1"
pytest-black = "^0.3.12"
pytest-asyncio = "^0.14.0"
jupyterlab = "^2.2.9"
ipykernel = "^5.3.4"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov --cov-report=xml:cov.xml --black"